backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
bcrypt==4.2.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.1.8
email-validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.116.1
h11==0.16.0
html5lib==1.1
httpx==0.27.2
httptools==0.6.4
idna==3.10
lxml==6.0.0
multitasking==0.0.12
numpy==2.0.2
pandas==2.3.1
passlib[bcrypt]==1.7.4
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.42
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
webencodings==0.5.1
websockets==15.0.1
yfinance==0.2.65
